backend/requirements.txt
fastapi==0.133.1
uvicorn==0.41.0

SQLAlchemy==2.0.47
psycopg2-binary==2.9.11

python-dotenv==1.2.1
python-multipart==0.0.22
python-jose==3.5.0
passlib==1.7.4
bcrypt==3.2.2
email-validator==2.3.0


numpy==2.4.2
huggingface_hub>=0.23.0

groq==1.0.0

python-docx==1.2.0
pypdf==6.7.4
